+++ src/main/kotlin/com/example/demo/entity/IncomeType.kt
@@ -6,6 +6,6 @@ data class IncomeType(val id: Int, val name: String)
 
 val incomeTypeMapper: (row: ResultSet, _: Any?) -> IncomeType = { row,_ ->
     IncomeType(
-        row.getInt("id") as Int, row.getString("name") as String
+        row.getInt("id"), row.getString("name") as String
     )
 }
